Define One-to-Many relationship between LiveGame and Tournament
This commit will define a bi-directional One-to-Many relationship between LiveGame and Tournament. 
A tournament will not have many live games, so a bi-directional relationship will not incur a high performance cost.
In addition, LiveGameConfig will be be removed, and all test data for tournaments, live games and pgns will be found in TournamentConfig.